requirements.txt
openai==1.70.0
streamlit==1.44.1
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=1.1.0
tqdm>=4.66.1
